fixed valid versions / token bug
